Remove deprecated DeprecatedPrepublishingViewController
