+++ WordPress/Classes/Services/Stories/StoryEditor.swift
@@ -212,14 +212,6 @@ class StoryEditor: CameraController {
 }
 
 extension StoryEditor: PublishingEditor {
-    var prepublishingIdentifiers: [PrepublishingIdentifier] {
-        if RemoteFeatureFlag.jetpackSocialImprovements.enabled() {
-            return  [.title, .visibility, .schedule, .tags, .categories, .autoSharing]
-        }
-
-        return  [.title, .visibility, .schedule, .tags, .categories]
-    }
-
     var prepublishingSourceView: UIView? {
         return nil
     }
